requirements.txt
streamlit
chainlit
azure-identity
azure-ai-inference
openai
openai-agents
google-genai
google-generativeai
python-dotenv
pydantic
semantic-kernel
mcp-agent
asyncio